BJS2-1414 task refactor.
refactor() {
 -+ SubscriptionControllerImpl was renamed as SubscriptionController and the test class as well
 -+ Moved validation to SubscriptionServiceImpl
}
